cherry picked: Refactor widget imports
(cherry picked from commit fc23e4e58ec0b2f3223fc94cdd18a17542257fa0)
